Move advice definition list to its own module
